backend/pyproject.toml
[dependency-groups]
dev = ["debugpy<2.0.0,>=1.8.1"]

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
authors = [{ name = "zipped ", email = "100anonyo@gmail.com" }]
license = { text = "MIT" }
requires-python = "<4.0,>=3.9"
dependencies = [
    "anyio<5.0.0,>=4.3.0",
    "beautifulsoup4<5.0.0,>=4.12.3",
    "bs4<1.0.0,>=0.0.2",
    "certifi<2025.0.0,>=2024.2.2",
    "charset-normalizer<4.0.0,>=3.3.2",
    "click<9.0.0,>=8.1.7",
    "colorama<1.0.0,>=0.4.6",
    "cython<4.0.0,>=3.0.9",
    "dnspython<3.0.0,>=2.6.1",
    "fastapi>=0.110.0,<1.0.0",
    "faust-cchardet<3.0.0,>=2.1.19",
    "gunicorn<22.0.0,>=21.2.0",
    "h11<1.0.0,>=0.14.0",
    "idna<4.0,>=3.6",
    "lxml<6.0.0,>=5.1.0",
    "meilisearch-python-sdk<3.0.0,>=2.9.0",
    "pydantic<3.0.0,>=2.6.4",
    "pymongo<5.0.0,>=4.6.2",
    "python-dotenv<2.0.0,>=1.0.1",
    "redis<6.0.0,>=5.0.3",
    "requests<3.0.0,>=2.31.0",
    "sniffio<2.0.0,>=1.3.1",
    "soupsieve<3.0,>=2.5",
    "tqdm<5.0.0,>=4.66.2",
    "typing-extensions<5.0.0,>=4.10.0",
    "urllib3<3.0.0,>=2.2.1",
    "uvicorn<1.0.0,>=0.28.0",
    "celery[redis]<6.0.0,>=5.4.0",
    "sentry-sdk[celery]<3.0.0,>=2.0.1",
    "meilisearch<1.0.0,>=0.31.4",
    "jinja2>=3.1.5",
]
name = "wallstreetlocal"
version = "2.0.0"
description = "Free and open-source stock tracking website for America's biggest money managers. Democratizing SEC filings by making them more accessible and useful."
